requirements.txt
Flask
Flask-SocketIO
google-cloud-speech
SpeechRecognition